-- Testes unitários [*] Mudança e comentário no teste ExpressaoCurrDuasCasasDecimais para explicar o comportamento do Win 64.
git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk2@18306 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Testes/FPCUnit/ACBrComum/acbrutiltest.pas
@@ -3296,9 +3296,17 @@ begin
   // 0.99 x 0.995 = 0,98505, porém "currTotal" é um currency, que somente usa 4 casas decimais, portanto será: 0,9850
   currVal := 0.98;
   CheckEquals(currVal , RoundABNT(currTotal, 2), 0.00001);
-  // RoundABNT tem um parâmetro do tipo "Double", portanto aqui todas as casas decimais serão utilizadas
+
   currVal := 0.99;
-  CheckEquals( currVal, RoundABNT(currValorUnit * currQtde, 2), 0.00001);
+  //   RoundABNT tem um parâmetro do tipo "Double" isso pode fazer com que expressões
+  // mudem o valor passado para a RoundABNT dependendo do compilador.
+  //   Por exemplo, O teste comentado abaixo falha quando o compilador é Win 64 bits mas passa no Win 32:
+  //CheckEquals( currVal, RoundABNT(currValorUnit * currQtde, 2), 0.00001);
+  //   O que acontece é que no Win 32, as variáveis são convertidas para Double antes da multiplicação e,
+  // assim, todas as casas decimais serão utilizadas. Já no Win 64,
+  // as variáveis são multiplicadas como Currency e só depois convertidas para Double.
+  //   No Win 64 o teste fica como abaixo:
+  CheckEquals( currVal, RoundABNT((Double(currValorUnit * Double(currQtde))), 2), 0.00001);
 end;
 
 procedure RoundABNTTest.ExpressaoDblDuasCasasDecimais;
